UseCase parcial para criar um usuario com roles
Fix: Relation User -> Pivot -> Role para User -> Role sem passar por pivot

+++ src/infrastructure/entities/user_roles_pivot.rs
@@ -6,8 +6,8 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "user_roles_pivot")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub role_id: i64,
-    pub user_id: i64,
+    pub role_id: i32,
+    pub user_id: i32,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
